♻️ move components to folders
