+++ packages/common-ui-web/src/components/VehicleDynamicWireframe/VehicleDynamicWireframe.style.ts
@@ -1,0 +1,11 @@
+import { Styles } from '@monkvision/types';
+
+export const styles: Styles = {
+  notCarPart: {
+    display: 'none',
+  },
+  selectAble: {
+    pointerEvents: 'fill',
+    cursor: 'pointer',
+  },
+};
