+++ packages/common-ui-web/src/components/VehicleDynamicWireframe/index.ts
@@ -1,0 +1,1 @@
+export * from './VehicleDynamicWireframe';
