+++ packages/common-ui-web/src/components/VehiclePartSelection/VehiclePartSelection.style.ts
@@ -1,0 +1,12 @@
+import { Styles } from '@monkvision/types';
+
+export const styles: Styles = {
+  wrapper: {
+    position: 'absolute',
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    inset: '0 0 0 0',
+    gap: '30px',
+  },
+};
