+++ packages/common-ui-web/src/components/VehiclePartSelection/index.ts
@@ -1,0 +1,1 @@
+export * from './VehiclePartSelection';
